requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.4
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
